Genuine app icon
When I drew the first version of the icon, I merely modified the icon
from Apple’s Keychain Access.app, to make it fit in as another OS X
utility. With a completely self-drawn icon however, we’re able to both
move away from blatant copying and also express what signifies the
inner works of PGP (that of keypairs).